+++ src/messageBag.js
@@ -16,7 +16,9 @@ MessageBag.prototype.set = require('./messages/set.js');
 MessageBag.prototype.list = require('./messages/list.js');
 MessageBag.prototype.macro = require('./messages/macro.js');
 MessageBag.prototype.forget = require('./messages/forget.js');
-MessageBag.prototype.forceMacro = require('./messages/macro.js');
+MessageBag.prototype.forceMacro = require('./messages/forceMacro.js');
+MessageBag.prototype.localMacro = require('./messages/localMacro.js');
+MessageBag.prototype.forceLocalMacro = require('./messages/forceLocalMacro.js');
 
 const MessageBagFactory = function (validator = {}) {
 	return new MessageBag(validator);